Rename ninja-jaxrs-routes to ninja-jaxy-routes to avoid confusion
