+++ ninja-jaxy-routes/src/test/resources/conf/messages.properties
