+++ ninja-jaxy-routes/src/test/resources/ehcache.xml
